+++ heroku-api/src/main/java/com/heroku/api/App.java
@@ -1,7 +1,6 @@
 package com.heroku.api;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 /**
  * Data model for a Heroku App. Also serves as a builder class when making requests to create an app.
@@ -25,8 +24,8 @@ public class App implements Serializable {
     String git_url;
     String buildpack_provided_description;
     String released_at;
-    BigInteger slug_size;
-	  BigInteger repo_size;
+    long slug_size;
+	  long repo_size;
     int dynos;
 	  int workers;
 
@@ -100,11 +99,11 @@ public class App implements Serializable {
         this.buildpack_provided_description = buildpack_provided_description;
     }
 
-    private void setSlug_size(BigInteger slug_size) {
+    private void setSlug_size(long slug_size) {
         this.slug_size = slug_size;
     }
 
-    private void setRepo_size(BigInteger repo_size) {
+    private void setRepo_size(long repo_size) {
         this.repo_size = repo_size;
     }
 
@@ -167,11 +166,11 @@ public class App implements Serializable {
         return repo_migrate_status;
     }
 
-    public BigInteger getSlugSize() {
+    public long getSlugSize() {
         return slug_size;
     }
 
-    public BigInteger getRepoSize() {
+    public long getRepoSize() {
         return repo_size;
     }
